desafio2_python/requirements.txt
pandas
langchain
langchain-openai
langchain-experimental
python-dotenv
tabulate